+++ Src/FluentAssertions/Common/Clock.cs
@@ -25,16 +25,4 @@ namespace FluentAssertions.Common
 
         public ITimer StartTimer() => new StopwatchTimer();
     }
-
-    internal class StopwatchTimer : ITimer
-    {
-        private readonly Stopwatch stopwatch;
-
-        public StopwatchTimer()
-        {
-            stopwatch = Stopwatch.StartNew();
-        }
-
-        public TimeSpan Elapsed => stopwatch.Elapsed;
-    }
 }
