+++ Tests/FluentAssertions.Specs/TestTimer.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using FluentAssertions.Common;
+
+namespace FluentAssertions.Specs
+{
+    internal class TestTimer : ITimer
+    {
+        private readonly Func<TimeSpan> getElapsed;
+
+        public TestTimer(Func<TimeSpan> getElapsed)
+        {
+            this.getElapsed = getElapsed;
+        }
+
+        public TimeSpan Elapsed => getElapsed();
+    }
+}
